ADHOC-FIX (update config values) Change behavior to insert or update
As discovered on showroom deployment, the config value for the base_url
was not set. This results in an showroom where no links are working in
Magento. My assumption in using this is that when I am setting magento
core configuration data via `magento_core_config_data` variable from
ansible, is that this values are set after the deployment.

Actualy this was only the case if this setting is already existing in
the database. Changing to an `insert` statement with `ON DUPLICATE KEY
UPDATE` is an esy way to achive this behavior. As there is an
'UNIQUE KEY `CORE_CONFIG_DATA_SCOPE_SCOPE_ID_PATH` (`scope`,`scope_id`,`path`)'
for the `core_config_data` table.