ADHOC fix (cron jobs): better cron-job setup
to get things done properly we need to disable message consumers triggering via
cron and instead call them manually with individual mutexes (we use flock).
Additionally there is a RabbitMQ installed on the server that will be working
as message broker if magento configuration will allow to use it.

See also
https://magento.stackexchange.com/questions/285385/magento-2-3-2-ce-and-cron-jobs-flock-and-rabbitmq