ADHOC chore (cron jobs): better naming
added explicit name to the operation. The idea here to see the list of
consumers that we are then installing on magento, it might be that we  won't
actually install them (depending on config), but we are still checking them.

Actually this change might break with magento 2.2.x or earlier, cause that
cli command might simply be not yet included in magento