ADHOC feat (cleaup): auto-cleanup
automatic cleanup of older clones to keep the disk usage under control. Those
clones/releases are usually around 1GB in size, having 30 of them on the
machine is easy within a week if project is atively developed. That means that
most of this space is pretty much wasted, cause all those clones are anyways
stored in AWS/Azure we don't need to keep them on the machine for backup
purposes or whatsoever. We only need few recent ones to ensure fast roll-back
in case of failed deploy.

The roll back is expected to be performed manually as of now. There is no
automation prepared or planned for it at the moment.

Using `shell` and not `command` here cause `command` is not able to interpret
`\(` correctly and pass that to the actual shell. It is an element of `find`
parameter syntax, we are using a construct there to do a negation.

The actual one-liner command was borrowed from other project where release
activation/deployment is done with a shell script, so the command itself is
battle-proven.