ADHOC feat (cleaup): auto-cleanup
automatic cleanup of older clones to keep the disk usage under control. Those
clones/releases are usually around 1GB in size, having 30 of them on the
machine is easy within a week if project is atively developed. That means that
most of this space is pretty much wasted, cause all those clones are anyways
stored in AWS/Azure we don't need to keep them on the machine for backup
purposes or whatsoever. We only need few recent ones to ensure fast roll-back
in case of failed deploy.

The roll back is expected to be performed manually as of now. There is no
automation prepared or planned for it at the moment.

This is yet another implementation for the problem explained above.
This implementaion is also backported from another project, so it is also
battle-prven so to say, at the same time it ansible native which makes it a bit
more readable. Author leaves both implementation stay in the history and not
squash/rebase to show possible alternative if later on additional constainst
will be discovered.

Using ctime as the actual file creation date and also non-fakeable attribute.

    http://lists.gnu.org/archive/html/coreutils/2010-08/msg00010.html
    ctime cannot be faked (at least it's not intended to be fakeable):

    POSIX says that atime and mtime are user-settable to arbitrary times via
    the utimensat() family of syscalls, but that ctime must unfakeably track
    the current time of any action that changes a file's metadata or contents.