AdHoc (change) Use relative pathes for the shared folders
It was discovered that deployment was failing while the server was under
load. The reason was, that magento was recreating the shared folders
between the deletion and the symlinking.

We move this step now to be done before the symlink is moved to the new
magento deployed folder. This is also saving some downtime in the
deployment.

*#####################
*## Breaking Change ##
*#####################

This includes a backwards incompatible change. As the targets for the
shared folders needs now to be defined relative to the application root
folder. Absolute pathes are not possible anymore.