ADHOC feat redis-cache: force expiry for (almost) all cache entries
we had a problem since a while that magento was not setting TTL on cache
entries in Redis, on some systems that results in Redis cache database
growing to huge numbers, like 16GB and starting to hit container/server
limitations.

After going with xdebung through magento core it became clear that
magento actually has a feature to control that, i.e. to force the
expiry.

See

    Cm_Cache_Backend_Redis::save()
    Cm_Cache_Backend_Redis::_getAutoExpiringLifetime()
    https://magento.stackexchange.com/questions/299854/redis-keys-are-keep-increasing-expiry-is-not-set-to-keys
    https://jmaitrehenry.ca/2017/11/22/found-keys-without-expiration-in-redis/

setting conditional breakpoints with xDebug we are always saving to
redis with expiration now, still we have ~10% of keys that don't have
TTL in Redis with those settings applied. At this moment I can not
explain that. Before this, something like 90% of keys in page_cache and
default cache backends were set without expiration.

Interesting enough, session cache backend is a different class and it
handles expiration better - there we always have 100% of keys with TTL
set.